refactor: move const field to graphql types module
