+++ src/blueprint/from_config/graphql/types/mod.rs
@@ -1,3 +1,5 @@
 mod http;
+mod const_field;
 
 pub use http::*;
+pub use const_field::*;
